squashing warnings - created 2.12 and 2.13 impls of EffectSet - put back "-Xfatal-warnings", - removed use of right projections in tests
