+++ diode-core/shared/src/test/scala/diode/util/RetryTests.scala
@@ -29,13 +29,13 @@ object RetryTests extends TestSuite {
       assert(r.isRight)
       val now = System.currentTimeMillis()
       // check that effect happens in the future
-      r.right.get._2.toFuture.flatMap { n =>
+      r.getOrElse((policy, effect(policy)))._2.toFuture.flatMap { n =>
         println(s"First retry after ${System.currentTimeMillis() - now}")
         assert(System.currentTimeMillis() - now > 150)
         // next retry
-        val rr = r.right.get._1.retry(new Exception, effect)
+        val rr = r.getOrElse((policy, effect(policy)))._1.retry(new Exception, effect)
         // check that effect happens in the future
-        rr.right.get._2.toFuture.map { n =>
+        rr.getOrElse((policy, effect(policy)))._2.toFuture.map { n =>
           println(s"Second retry after ${System.currentTimeMillis() - now}")
           assert(System.currentTimeMillis() - now > 550)
         }
